aois_3

+++ 2 year/AOIS/src/aois_3/aois_3/BidirectionalNN.h
@@ -6,15 +6,15 @@ class BidirectionalNN
 {
 private:
 	int n, m;
-	std::vector<std::vector<int>> Weight;
-	std::vector<std::vector<int>> WeightTransposed;
+	std::vector<std::vector<double>> Weight;
+	std::vector<std::vector<double>> WeightTransposed;
 
-	std::vector<std::vector<int>> multiplyMatrices(const std::vector<std::vector<int>>& matrix1, const std::vector<std::vector<int>>& matrix2);
-	std::vector<std::vector<int>> transposeMatrix(std::vector<std::vector<int>>);
-	std::vector<std::vector<int>> roundVector(std::vector<std::vector<int>>);
+	std::vector<std::vector<double>> multiplyMatrices(const std::vector<std::vector<double>>& matrix1, const std::vector<std::vector<double>>& matrix2);
+	std::vector<std::vector<double>> transposeMatrix(std::vector<std::vector<double>>);
+	std::vector<std::vector<double>> roundVector(std::vector<std::vector<double>>);
 public:
 	BidirectionalNN(int n, int m);
-	std::vector<std::vector<int>> function(bool mode, std::vector<std::vector<int>>);
-	void initializeWeight(std::vector<std::vector<int>> X, std:: vector<std::vector<int>> Y);
+	std::vector<std::vector<double>> function(bool mode, std::vector<std::vector<double>>);
+	void initializeWeight(std::vector<std::vector<double>> X, std:: vector<std::vector<double>> Y);
 };
